Move EnumHelper to Avalonia.Base
