+++ src/Avalonia.Base/Avalonia.Base.csproj
@@ -7,7 +7,6 @@
   </PropertyGroup>
   <ItemGroup>
     <EmbeddedResource Include="Assets\*.trie" />
-    <Compile Include="..\Shared\EnumHelper.cs" />
   </ItemGroup>
   <Import Project="..\..\build\Base.props" />
   <Import Project="..\..\build\Binding.props" />
